% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network_effects.R
\name{get_top_k_gene_indices}
\alias{get_top_k_gene_indices}
\title{Get indices of the top-k correlated genes (including if exist target gene(s))}
\usage{
get_top_k_gene_indices(Sigma, genes, target, k)
}
\arguments{
\item{Sigma}{Covariance (or correlation) matrix.}

\item{genes}{Character vector of gene names (length = ncol(Sigma)).}

\item{target}{Gene index(es) or name(s).}

\item{k}{Number of top genes to return (including the target).}
}
\value{
Integer vector of indices for the top-k genes.
}
\description{
Get indices of the top-k correlated genes (including if exist target gene(s))
}
\keyword{internal}
