% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network_effects.R
\name{new_GLgraph}
\alias{new_GLgraph}
\title{Create a GLgraph object}
\usage{
new_GLgraph(fit, k, target)
}
\arguments{
\item{fit}{A list from run_glasso()/fit_glasso()/fit_glasso_raw().
Must contain: $sigma, $omega, $features.}

\item{k}{Integer. Number of genes to include.}

\item{target}{Gene name(s) or index(es).}
}
\value{
An object of class "GLgraph".
}
\description{
Create a GLgraph object
}
