% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network_effects.R
\name{new_GLnode}
\alias{new_GLnode}
\title{Create a GLnode object}
\usage{
new_GLnode(index, neighbours = integer(0), knocked = FALSE)
}
\arguments{
\item{index}{Integer. The gene index for this node.}

\item{neighbours}{Integer vector of neighbor gene indices.}

\item{knocked}{Logical. Whether the gene is knocked or not.}
}
\value{
An object of class "GLnode".
}
\description{
Create a GLnode object
}
