% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{pbmc_small}
\alias{pbmc_small}
\title{pbmc_small}
\format{
A Seurat object.
}
\source{
Seurat vignettes sample dataset
}
\usage{
pbmc_small
}
\description{
A tiny Seurat read10x readble single cell counting matrix for examples and tests.
}
\references{
Hao, Y. et al. (2021). Integrated analysis of multimodal single-cell data.
\emph{Cell} 184(13):3573–3587.
}
\keyword{datasets}
