% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network_effects.R
\name{visnetwork_from_GLgraph}
\alias{visnetwork_from_GLgraph}
\title{Convert GLgraph object to a visNetwork graph}
\usage{
visnetwork_from_GLgraph(
  G,
  color_normal = "lightblue",
  color_knocked = "salmon"
)
}
\arguments{
\item{G}{A GLgraph object.}

\item{color_normal}{Color for unknocked nodes.}

\item{color_knocked}{Color for knocked nodes.}
}
\value{
A visNetwork object.
}
\description{
Convert GLgraph object to a visNetwork graph
}
