% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network_effects.R
\name{visnetwork_toggle_knock}
\alias{visnetwork_toggle_knock}
\title{Toggle knock state of a node and update visNetwork graph}
\usage{
visnetwork_toggle_knock(
  G,
  clicked,
  color_normal = "lightblue",
  color_knocked = "salmon"
)
}
\arguments{
\item{G}{A GLgraph object.}

\item{clicked}{Integer ID of clicked node.}

\item{color_normal}{Color for unknocked nodes.}

\item{color_knocked}{Color for knocked nodes.}
}
\value{
Updated visNetwork object.
}
\description{
Toggle knock state of a node and update visNetwork graph
}
