% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_glasso.R
\name{zscore_matrix_with_params}
\alias{zscore_matrix_with_params}
\title{Variable gene selection & Z-scoring for raw matrix}
\usage{
zscore_matrix_with_params(mat, nfeatures = 2000)
}
\arguments{
\item{mat}{normalized log1p matrix, cells × genes}

\item{nfeatures}{number of highly variable genes to select}
}
\value{
list(z, mu, sd, features)
}
\description{
Variable gene selection & Z-scoring for raw matrix
}
