% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_glasso.R
\name{zscore_seurat}
\alias{zscore_seurat}
\title{Z-score Seurat data (internal)}
\usage{
zscore_seurat(obj, nfeatures = 2000)
}
\arguments{
\item{obj}{Seurat object.}

\item{nfeatures}{number of variable features to select.}
}
\value{
numeric matrix (features x cells)
}
\description{
Selects variable features, scales them, and returns the scaled matrix.
Assumes the object is already normalized.
}
\keyword{internal}
