% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{pbmc_small}
\alias{pbmc_small}
\title{pbmc_small}
\format{
gene-by-cell UMI count matrix (dgCMatrix)RDS-compressed sparse matrix
}
\source{
Seurat https://satijalab.org/seurat/articles/pbmc3k_tutorial
}
\usage{
pbmc_small
}
\description{
A dataset of Peripheral Blood Mononuclear Cells (PBMC)
freely available from 10X Genomics.
There are 2,700 single cells that were sequenced on the Illumina NextSeq 500.
This is coming from the cellranger pipeline from 10X, returning a unique
molecular identified (UMI) count matrix.
For example dataset we take the first 1000 cells.
For detailed generating step, see data-raw/pbmc10x.R
}
\references{
Hao, Y. et al. (2021). Integrated analysis of multimodal single-cell data.
\emph{Cell} 184(13):3573–3587.
}
\keyword{datasets}
