% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_glasso.R
\name{preprocess_matrix_raw}
\alias{preprocess_matrix_raw}
\title{Quality control + normalization for a raw count matrix (cells × genes)}
\usage{
preprocess_matrix_raw(
  mat,
  min_genes = 200,
  max_genes = 2500,
  max_mt = 5,
  normalize = TRUE,
  sf = 10000
)
}
\arguments{
\item{mat}{numeric matrix, cells × genes (raw counts)}

\item{min_genes}{minimum detected genes per cell}

\item{max_genes}{maximum detected genes per cell}

\item{max_mt}{maximum mitochondrial percent (default 5)}

\item{normalize}{whether to normalize (default TRUE)}

\item{sf}{scale factor for normalization (default 10000)}
}
\value{
a normalized log1p matrix (cells × genes)
}
\description{
Quality control + normalization for a raw count matrix (cells × genes)
}
