% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_glasso.R
\name{preprocess_seurat_data}
\alias{preprocess_seurat_data}
\title{Quality control and normalization
for a given seurat single cell expression matrix
(Seurat Object)}
\usage{
preprocess_seurat_data(
  obj,
  min_genes = 200,
  max_genes = 2500,
  max_mt = 5,
  normalize = TRUE,
  sf = 10000
)
}
\arguments{
\item{obj}{Seurat object or matrix. If Seurat, QC is applied.}

\item{min_genes}{minimum genes per cell (default 200)}

\item{max_genes}{maximum genes per cell (default 2500)}

\item{max_mt}{maximum percent mitochondrial genes (default 5)}

\item{normalize}{boolean for optional normalization (default True)}

\item{sf}{scale.factor for normalization (defult 10000)}
}
\value{
subset of the seurat object
}
\description{
Quality control and normalization
for a given seurat single cell expression matrix
(Seurat Object)
}
